Implement scorer api for Upstash and refactor
